Recode 5-

+++ 0x04-more_functions_nested_loops/5-more_numbers.c
@@ -10,8 +10,8 @@
  */
 void more_numbers(void)
 {
-	int long c = 0;
-	int long count = 0;
+	int c = 0;
+	int count = 0;
 
 	while (count < 10)
 	{
